backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
openai==1.14.3
python-dotenv==1.0.1
pandas==1.5.3
beautifulsoup4
googlesearch-python
requests